Refactor: Modal 컴포넌트와 context 분리
