+++ FE/src/components/Modal/style.ts
