+++ FE/src/components/SearchBox/PriceChartModal/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import Modal from '@/components/Modals/Modal';
+import Modal from '@/components/Modal';
 import PriceChart from '@/components/SearchBox/PriceChartModal/PriceChart';
 import { PriceChartData, PriceChartDataProps } from '@/data/chartData';
